# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitzania",
        "description": "This is API documentation of OPEN API SWAGGER for developers to test API.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "fitzania@yopmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/Fitzania/api/",
            "description": "SwaggerHUB API Mocking"
        }
    ],
    "paths": {
        "/GymRegistration": {
            "post": {
                "tags": [
                    "Gym API Section"
                ],
                "summary": "Gym registration process goes here",
                "operationId": "Registration",
                "requestBody": {
                    "description": "Pass gym registration detail",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_type",
                                    "phone",
                                    "device_type"
                                ],
                                "properties": {
                                    "user_type": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "4785236956"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "example": "TestDeviceToken"
                                    },
                                    "firebase_uid": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/Registration": {
            "post": {
                "tags": [
                    "User API Section"
                ],
                "summary": "User registration process goes here",
                "description": "User Type : 1-IOS, 2-Android, 3-Website, 4-Admin",
                "operationId": "Registration",
                "requestBody": {
                    "description": "Pass user registration detail",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_type",
                                    "phone",
                                    "device_type"
                                ],
                                "properties": {
                                    "user_type": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "4785236956"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "example": "TestDeviceToken"
                                    },
                                    "firebase_uid": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}